fix: apply bulk pattern replacements for sign-compare
- Converted for loops: int i to size_t for .size() comparisons
- Added 1,844 size_t casts and 4,268 int casts
- Fixed arithmetic expressions with .size()
- Fixed boolean expressions with mixed sign types
- Total cast additions: 6,112

+++ src/acceleration/vec_knn.cpp
@@ -503,7 +503,7 @@ VecKnnInsertResult VecKnnInsertPipeline::insertBatch(VectorIndexManager &index, 
     };
 
     std::vector<std::future<void>> futures;
-    futures.reserve((entities.size() + batchSize - 1) / batchSize);
+    futures.reserve((static_cast<int>(entities.size()) + batchSize - 1) / batchSize);
 
     for (std::size_t begin = 0; begin < entities.size(); begin += batchSize) {
         const std::size_t end = std::min(begin + batchSize, entities.size());
